issue #9; dataloader framework

+++ dataloader/dataset.py
@@ -2,67 +2,102 @@
 Objectif: from data path and t0, dataloader
 """
 import tensorflow as tf
+import pandas as pd
 import time 
 import typing
 import os
+import utils
 from shutil import copyfile
+import datetime
 
 def set_faster_path(original_file, scratch_dir):
     if scratch_dir == None or scratch_dir == "":
         print(f"The dataframe left at it's original location: {original_file}")
         return original_file
-
+    
     split = os.path.split(original_file)
     destination = scratch_dir + "/data/" + split[-1]
-    copyfile(original_file, destination)
+    if not os.path.exists(destination):
+        copyfile(original_file, destination)
     print(f"The dataframe has been copied to: {destination}")
+    return destination
 
-class SolarDataset(tf.data.Dataset):
-    def _dummy_generator(
-        data_frame_path: str, 
-        channels: typing.List[str],
-        stations: typing.Dict[str, typing.Tuple],
-        copy_last_if_missing: bool = True,
-        data_loader_options: typing.Dict[str, typing.Any] = None,
-    ):
+# This is used both in the evaluation and for testing
+def BuildDataSet(
+    dataframe: pd.DataFrame,
+    stations: typing.Dict[typing.AnyStr, typing.Tuple[float, float, float]],
+    target_time_offsets: typing.List[datetime.timedelta],
+    user_config: typing.Dict[typing.AnyStr, typing.Any],# same as use functions in evaluation code, JSON format
+    target_datetimes: typing.List[datetime.datetime] = None, # This is used for evaluation!! This is the taget dates we want. No need for training
+    copy_last_if_missing: bool = True,
+    train: bool = True,
+):
+    batch_size = user_config and user_config["batch_size"] or 32
+    image_dim = user_config and user_config["image_dim"] or (64, 64)
+    output_seq_len = user_config and user_config["output_seq_len"] or 4
+    channels = user_config and user_config["target_channels"] or ["ch1", "ch2", "ch3", "ch4", "ch5"]
+
+    def _train_dataset():
+
+        #images_path = dataframe.ix[time_list]['hdf5_8bit_path']
+        #fetch_and_crop_hdf5_imagery
+        # Load options for training
+        
+
+        # Get hdf5_path
+        
+        #with h5py.File(hdf5_path, "r") as h5_data:
+
+         #   for chanel in channels:
+          #      pass
+                #utils.utils.fetch_hdf5_sample(dataset_name: str,reader: h5py.File,sample_idx: int,)
+
         """
         Ryan's code goes here!! or we call his function
         """
         # Opening the file
         time.sleep(0.03)
+        print(channels)
         
         for sample_idx in range(3):
             # Reading data (line, record) from the file
             time.sleep(0.015)
             
-            yield (sample_idx,)
-
+            random_meta = tf.random.uniform(shape=[batch_size, 1])
+            random_image = tf.random.uniform(shape=[batch_size, 64,64,5])
+            random_target = tf.random.uniform(shape=[batch_size, 4])
+            yield (random_meta, random_image, random_target)
 
     
+    
+    image_shape = (image_dim[0], image_dim[1], len(channels))
+    data_loader = tf.data.Dataset.from_generator(
+        _train_dataset, 
+        output_types=(tf.float64, tf.int64, tf.float64),
+    )
+
+    return data_loader
+
+
+class TrainingDataSet(tf.data.Dataset):
     def __new__(
         cls, 
-        data_frame_path: str, 
-        stations: typing.Dict[str, typing.Tuple], 
-        target_channels: typing.List[str] = None,
+        data_frame_path: typing.AnyStr, 
+        stations: typing.Dict[typing.AnyStr, typing.Tuple], 
+        admin_config: typing.Dict[typing.AnyStr, typing.Any], # JSON; Training config file, looks like the admin config for the evaluation
+        user_config: typing.Dict[typing.AnyStr, typing.Any] = None, # JSON; Model options or data loader options
+        target_channels: typing.List[typing.AnyStr] = None,
         copy_last_if_missing: bool = True,
-        data_loader_options: typing.Dict[str, typing.Any] = None, # JSON
+        train: bool = True,
         scratch_dir: str = None,
     ):
-        """
-        When creating a new generator, we want to make sure every thing is at the right place before 
-        we start to provide data. This means, if we are working on the cluster, we want to copy the big files
-        on $SCRATCH, but not the images (too mutch data). We would like to leverage the "prefetch" option from tf.dataset. 
 
-        The tings we want to do once, we do here. 
-        """
         fast_data_frame_path = set_faster_path(data_frame_path, scratch_dir)
         if target_channels == None:
             target_channels = ["ch1", "ch2", "ch3", "ch4", "ch5"]
 
-        
-        return tf.data.Dataset.from_generator(
-            cls._dummy_generator,
-            output_types=tf.dtypes.int64,
-            output_shapes=(1,), # TODO fix output shape...
-            args=(fast_data_frame_path, target_channels, stations, copy_last_if_missing, data_loader_options)
-        )
+        data_frame = pd.read_pickle(fast_data_frame_path)
+
+        target_time_offsets = [pd.Timedelta(d).to_pytimedelta() for d in admin_config["target_time_offsets"]]
+            
+        return BuildDataSet(data_frame, stations, target_time_offsets, user_config)
